misc/HtmlTestSuite/results/test-0077.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml">
<head>
<meta charset="utf-8"/>
<title>Baseline of inferred mrows</title>
<link href="https://w3c.github.io/mathml-core/#horizontally-group-sub-expressions-mrow" rel="help"/>
<link href="https://w3c.github.io/mathml-core/#radicals-msqrt-mroot" rel="help"/>
<link href="https://w3c.github.io/mathml-core/#style-change-mstyle" rel="help"/>
<link href="https://w3c.github.io/mathml-core/#error-message-merror" rel="help"/>
<link href="https://w3c.github.io/mathml-core/#making-sub-expressions-invisible-mphantom" rel="help"/>
<link href="https://w3c.github.io/mathml-core/#the-top-level-math-element" rel="help"/>
<link href="https://www.w3.org/TR/MathML3/chapter3.html#presm.menclose" rel="help"/>
<link href="https://w3c.github.io/mathml-core/#adjust-space-around-content-mpadded" rel="help"/>
<meta content="Baseline for mrow-like elements is correct." name="assert"/>
<script src="/mathml/support/feature-detection.js"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script type="text/javascript">
  setup({ explicit_done: true });
  window.addEventListener("load", runTests);
  function runTests()
  {
      ["Mrow", "Sqrt", "Style", "Error", "Phantom", "Math", "Menclose", "Mpadded", "Unknown", "Mtd", "None", "Mprescripts"].forEach((tag) =&gt; {
          var x = document.getElementById(`above${tag}`).getBoundingClientRect();
          var y = document.getElementById(`below${tag}`).getBoundingClientRect();
          test(function() {
              assert_true(MathMLFeatureDetection.has_mspace());
              assert_equals(x.bottom, y.top);
          }, `baseline alignment inside ${tag}`);
      });
      done();
  }
</script>
</head>
<body>
  <div id="log"/>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML"><mrow><mspace height="30px" id="aboveMrow" style="background: purple" width="10px"/><mspace depth="30px" id="belowMrow" style="background: blue" width="10px"/></mrow></math>
    <math xmlns="http://www.w3.org/1998/Math/MathML"><msqrt><mspace height="30px" id="aboveSqrt" style="background: purple" width="10px"/><mspace depth="30px" id="belowSqrt" style="background: blue" width="10px"/></msqrt></math>
    <math xmlns="http://www.w3.org/1998/Math/MathML"><mstyle><mspace height="30px" id="aboveStyle" style="background: purple" width="10px"/><mspace depth="30px" id="belowStyle" style="background: blue" width="10px"/></mstyle></math>
    <math xmlns="http://www.w3.org/1998/Math/MathML"><merror><mspace height="30px" id="aboveError" style="background: purple" width="10px"/><mspace depth="30px" id="belowError" style="background: blue" width="10px"/></merror></math>
    <math xmlns="http://www.w3.org/1998/Math/MathML"><mphantom><mspace height="30px" id="abovePhantom" style="visibility: visible;" width="10px"/><mspace depth="30px" id="belowPhantom" style="visibility: visible;" width="10px"/></mphantom></math>
    <math xmlns="http://www.w3.org/1998/Math/MathML"><mspace height="30px" id="aboveMath" style="background: purple" width="10px"/><mspace depth="30px" id="belowMath" style="background: blue" width="10px"/></math>
    <!-- menclose is treated as <unknown> in MathML Core -->
    <math xmlns="http://www.w3.org/1998/Math/MathML"><menclose notation="box"><mspace height="30px" id="aboveMenclose" style="background: purple" width="10px"/><mspace depth="30px" id="belowMenclose" style="background: blue" width="10px"/></menclose></math>
    <math xmlns="http://www.w3.org/1998/Math/MathML"><mpadded lspace="10px"><mspace height="30px" id="aboveMpadded" style="background: purple" width="10px"/><mspace depth="30px" id="belowMpadded" style="background: blue" width="10px"/></mpadded></math>
    <math xmlns="http://www.w3.org/1998/Math/MathML"><unknown><mspace height="30px" id="aboveUnknown" style="background: purple" width="10px"/><mspace depth="30px" id="belowUnknown" style="background: blue" width="10px"/></unknown></math>
    <math xmlns="http://www.w3.org/1998/Math/MathML"><mtable><mtr><mtd><mspace height="30px" id="aboveMtd" style="background: purple" width="10px"/><mspace depth="30px" id="belowMtd" style="background: blue" width="10px"/></mtd></mtr></mtable></math>
    <math xmlns="http://www.w3.org/1998/Math/MathML"><none><mspace height="30px" id="aboveNone" style="background: purple" width="10px"/><mspace depth="30px" id="belowNone" style="background: blue" width="10px"/></none></math>
    <math xmlns="http://www.w3.org/1998/Math/MathML"><mprescripts><mspace height="30px" id="aboveMprescripts" style="background: purple" width="10px"/><mspace depth="30px" id="belowMprescripts" style="background: blue" width="10px"/></mprescripts></math>
  </p>
</body>
</html>
